[1.0.2] build error fixed

+++ Gamnet/Library/Exception.h
@@ -63,7 +63,7 @@ private :
 
 }
 
-#define GAMNET_ERRNO(error) error, "ERR [", __FILE__, ":", __func__, "@" , __LINE__, "] ", #error, " "
+#define GAMNET_ERRNO(error) (int)error, "ERR [", __FILE__, ":", __func__, "@" , __LINE__, "] ", #error, " "
 #define GAMNET_ERRSTR(error) #error, "(", (int)error, ") "
 
 #endif /* EXCEPTION_H_ */
